# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Portway Control Plane API
  version: 1.0.0
  description: >-
    Control-plane API derived from docs/API.md. Phase 9 implements authentication,
    scoped projects/tunnels, configured relay metadata and credential issuance in
    PostgreSQL by default (explicit bounded memory mode is available for tests).
    Durable state and revocation survive API restart. Phase 12 implements scoped domains
    and DNS ownership verification. Phase 14 implements bounded scoped request logs,
    tunnel observations and operator Prometheus metrics. Phase 13 consumes
    these contracts through a browser dashboard adapter described in API.md; its
    browser-local routes are outside this control-plane base.
    Database admission/transaction failures return 503 STORAGE_UNAVAILABLE. Phase 11
    requires fresh HEALTHY Redis reports and capacity headroom for relay assignments;
    Redis failure returns 503 PRESENCE_UNAVAILABLE. Relay-scoped report/drain routes
    fence stale incarnations and keep application traffic local to its assigned relay.
    JSON bodies are at most 64 KiB. Global 429/503/413 error envelopes apply.
    Secret-bearing responses use Cache-Control no-store and are not replay-cached.
    This contract manages metadata; it does not carry public application traffic.
servers:
  - url: https://api.portway.example.com/api/v1
  - url: http://localhost:8080/api/v1
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '200':
          $ref: '#/components/responses/SessionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /auth/logout:
    post:
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '204':
          description: Session revoked.
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /me:
    get:
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '200':
          $ref: '#/components/responses/MeResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /projects:
    get:
      tags: [Projects]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '200':
          $ref: '#/components/responses/ProjectListResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Projects]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '201':
          $ref: '#/components/responses/ProjectResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      security:
        - bearerAuth: []
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '200':
          $ref: '#/components/responses/ProjectResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Projects]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '204':
          description: Project deleted according to lifecycle policy.
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /tunnels:
    get:
      tags: [Tunnels]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: projectId
          schema: { type: string }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/TunnelStatus' }
        - in: query
          name: relayId
          schema: { type: string }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '200':
          $ref: '#/components/responses/TunnelListResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Tunnels]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTunnelRequest'
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '201':
          $ref: '#/components/responses/TunnelResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /tunnels/{tunnelId}:
    parameters:
      - $ref: '#/components/parameters/TunnelId'
    get:
      tags: [Tunnels]
      security:
        - bearerAuth: []
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '200':
          $ref: '#/components/responses/TunnelResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Tunnels]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '204':
          description: Tunnel deleted or transitioned to its terminal lifecycle state.
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /tunnels/{tunnelId}/revoke:
    post:
      tags: [Tunnels]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TunnelId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '200':
          $ref: '#/components/responses/TunnelResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /tunnels/{tunnelId}/connect:
    post:
      tags: [Tunnels]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TunnelId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConnectRequest' }
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '200':
          $ref: '#/components/responses/ConnectResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
  /domains:
    get:
      tags:
        - Domains
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - in: query
          name: tunnelId
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/DomainListResponse'
        default: &id001
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags:
        - Domains
      description: Reserve a scoped canonical hostname and return its one-time 24h DNS TXT proof. Keyed retries do not replay proof.
      security:
        - bearerAuth: []
      responses:
        default: *id001
        '201':
          $ref: '#/components/responses/DomainChallengeResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDomainRequest'
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
  /domains/{domainId}:
    parameters:
      - $ref: '#/components/parameters/DomainId'
    get:
      tags:
        - Domains
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/DomainResponse'
        default: *id001
    delete:
      tags:
        - Domains
      description: Disable alias; retain hostname reservation and audit.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema: &id002
              type: object
              additionalProperties: false
      responses:
        '204':
          description: Domain disabled.
        default: *id001
...
  /relays:
    get:
      tags: [Relays]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '200':
          $ref: '#/components/responses/RelayListResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /relays/{relayId}:
    parameters:
      - $ref: '#/components/parameters/RelayId'
    get:
      tags: [Relays]
      security:
        - bearerAuth: []
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '200':
          $ref: '#/components/responses/RelayResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /tunnels/{tunnelId}/logs:
    get:
      tags: [Logs]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TunnelId'
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 4, default: 4 }
      responses:
        '200':
          $ref: '#/components/responses/LogListResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /internal/credentials/verify:
    post:
      tags: [Auth]
      description: Relay AUTH-only verification; never used by public request forwarding. Relay credential is scoped to relayId; raw tunnel tokens are not sent here. Superseded relay/generation assignments return AUTH_REVOKED; admitted local leases remain usable until replacement or expiry.
      security:
        - relayAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [tokenHash, relayId]
              properties:
                tokenHash: { type: string, pattern: '^[a-f0-9]{64}$' }
                relayId: { type: string }
      responses:
        default:
          $ref: '#/components/responses/ErrorResponse'
        '200':
          description: Verified tunnel and generation binding.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [tunnelId, generation, expiresAt]
                        properties:
                          tunnelId: { type: string }
                          generation:
                            { type: string, pattern: '^[1-9][0-9]{0,19}$' }
                          expiresAt: { type: string, format: date-time }
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    relayAuth:
      type: http
      scheme: bearer
      description: Separate operator-provisioned relay verification credential.
  parameters:
    ProjectId:
      in: path
      name: projectId
      required: true
      schema: { type: string }
...
    RelayId:
      in: path
      name: relayId
      required: true
      schema: { type: string }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 255
    Cursor:
      in: query
      name: cursor
      required: false
      schema: { type: string }
    Limit:
      in: query
      name: limit
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
  schemas:
...
    CreateProjectRequest:
      type: object
      additionalProperties: false
      required: [name, slug]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        slug:
          { type: string, maxLength: 63, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token: { type: string, minLength: 32, maxLength: 512, writeOnly: true }
      description: Phase 9 exchanges a provisioned API key for a one-hour bounded session. OAuth/device flow remains later work.
...
  responses:
...
    ErrorResponse:
      description: API error.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data: { type: 'null' }
                  error: { $ref: '#/components/schemas/ApiError' }
    SessionResponse:
      description: Authenticated session.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data:
                    type: object
                    properties:
                      session: { $ref: '#/components/schemas/Session' }
    MeResponse:
      description: Current user, organization and membership role.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data:
                    type: object
                    properties:
                      user: { $ref: '#/components/schemas/User' }
                      organization:
                        type: object
                        required: [id, name, slug]
                        properties:
                          id: { type: string }
                          name: { type: string }
                          slug: { type: string }
                      role:
                        { type: string, enum: [OWNER, ADMIN, MEMBER, VIEWER] }
    ProjectResponse:
      description: Project response.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data:
                    type: object
                    properties:
                      project: { $ref: '#/components/schemas/Project' }
    ProjectListResponse:
      description: Project list.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data:
                    type: object
                    properties:
                      projects:
                        type: array
                        items: { $ref: '#/components/schemas/Project' }
                  meta:
                    allOf:
                      - $ref: '#/components/schemas/Envelope/properties/meta'
...